Move utils to client class
